+++ tubular/numeric.py
@@ -21,7 +21,7 @@ from tubular.mixins import (
 )
 
 
-class BaseNumericTransformer(BaseTransformer):
+class BaseNumericTransformer(BaseTransformer, CheckNumericMixin):
     """
     Extends BaseTransformer for datetime scenarios.
 
@@ -48,32 +48,6 @@ class BaseNumericTransformer(BaseTransformer):
     def __init__(self, columns: list[str], **kwargs: dict[str, bool]) -> None:
         super().__init__(columns=columns, **kwargs)
 
-    def _check_numeric(self, X: pd.DataFrame) -> None:
-        """Raise a type error if a column to be operated on is not numeric
-
-        Parameters
-        ----------
-
-        X: pd.DataFrame
-            Data to validate
-
-        """
-
-        numeric_column_types = X[self.columns].apply(
-            pd.api.types.is_numeric_dtype,
-            axis=0,
-        )
-
-        if not numeric_column_types.all():
-            non_numeric_columns = list(
-                numeric_column_types.loc[~numeric_column_types].index,
-            )
-
-            msg = f"{self.classname()}: The following columns are not numeric in X; {non_numeric_columns}"
-            raise TypeError(msg)
-
-        return X
-
     def fit(
         self,
         X: pd.DataFrame,
@@ -93,7 +67,7 @@ class BaseNumericTransformer(BaseTransformer):
 
         super().fit(X, y)
 
-        self._check_numeric(X)
+        CheckNumericMixin.check_numeric_columns(self, X)
 
         return self
 
@@ -114,7 +88,7 @@ class BaseNumericTransformer(BaseTransformer):
 
         X = super().transform(X)
 
-        self._check_numeric(X)
+        CheckNumericMixin.check_numeric_columns(self, X)
 
         return X
 
@@ -337,6 +311,7 @@ class TwoColumnOperatorTransformer(
     NewColumnNameMixin,
     TwoColumnMixin,
     DataFrameMethodTransformer,
+    BaseNumericTransformer,
 ):
     """This transformer applies a pandas.DataFrame method to two columns (add, sub, mul, div, mod, pow).
 
@@ -437,14 +412,9 @@ class TwoColumnOperatorTransformer(
         -------
             pd.DataFrame: Input X with an additional column.
         """
-        # call DataFrameMethodTransformer.transform
+        # call appropriate parent transforms
         X = super(DataFrameMethodTransformer, self).transform(X)
-
-        is_numeric = X[self.columns].apply(pd.api.types.is_numeric_dtype, axis=0)
-
-        if not is_numeric.all():
-            msg = f"{self.classname()}: input columns in X must contain only numeric values"
-            raise TypeError(msg)
+        X = super(BaseNumericTransformer, self).transform(X)
 
         X[self.new_column_name] = getattr(X[[self.column1_name]], self.pd_method_name)(
             X[self.column2_name],
@@ -716,7 +686,7 @@ class InteractionTransformer(BaseNumericTransformer):
         return X
 
 
-class PCATransformer(CheckNumericMixin, BaseTransformer):
+class PCATransformer(BaseNumericTransformer):
     """Transformer that generates variables using Principal component analysis (PCA).
     Linear dimensionality reduction using Singular Value Decomposition of the
     data to project it to a lower dimensional space.